+++ src/lib/djerba/extract/constants.py
@@ -2,10 +2,8 @@
 
 AUTHOR = 'author'
 ASSAY_TYPE = 'assay_type'
-ASSAY_NAME = 'assay_name'
 COVERAGE = 'coverage'
 FAILED = 'failed'
 ONCOTREE_CODE = 'oncotree_code'
 PURITY_FAILURE = 'purity_failure'
-STUDY = 'study'
-REQ_ID = 'requisition_id'
+PROJECT = 'Project'